Improvd TestRemoveItemFromSessionView.available
